About-Myself
Hi everyone!
I signed up to GitHub because I am new to programming and I believe that it is a great place to start. I am still in school at this time, and I would like to start getting my feet wet before I graduate. I am currently working in manufacturing (operations/production/logistics, at the management level), and have been doing it for over 5 years. I am also multilingual, and I hope my background will help me succeed in this field. I am thinking of moving towards 'automation', as it relates to what I presently do, but this is "not set in stone", is just the thought of using the knowledge I already have. I wish everyone on this platform the best of success!!!